[415137] h2 model db should be indexed
Add index for the (PATH, CONTAINER_ID) columns of the FILE table. This
improves significantly the performance of executing select/update/delete
statements on the h2 db when it grows with many file records.

Other columns used in prepared statements are indexed automatically by
the h2 db and it is not necessary to explicitly create indices for them:
  - CONTAINERS > PATH - because it is UNIQUE
  - FILES > CONTAINER_ID - because it is a foreign key

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=415137
Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>