Bug 478947 - Allow create .sh file outside shell project
Two problems for achieving it spotted in DLTK Core:
* ScriptModuleUtil.getFileExtensions documents being fine to return null
but doesn't handle the case where languageToolkit might be null which
results in NPE instead of returning null for file extension.
* NewSourceModulePage.getFileName adds dot at the end of the file if it
wasn't able to find any extension in either language toolkit or proposed
file name. As it is common use case to have shell scripts without
extension method is changed to not append "." (dot) at the end of the
file name if extension is empty.

Change-Id: I1926da8782b01da09e24169bab735cfd441a9c4d
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>